Product Registration and Update done

+++ php_deleteUser.php
@@ -3,6 +3,9 @@
         include_once "dbConnection.php";
         $conn = dbConncetion();
 
+
+if(isset($_GET['id'])){
+        $id = $_GET['id'];
         $sql = "DELETE FROM customerdetails WHERE id = '$id'";
         if($conn->query($sql)){
             $_SESSION['msg'] = "User deleted successfully";
